pania_gold/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
django-jalali==7.2.0
jalali_core==1.0.0
jdatetime==5.0.0
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.2
tzdata==2024.2
